refactor(types): deduplicate
